add the ofcontroller module for implementing OpenFlow controllers
Define classes OpenflowControllerStub and
OpenflowControllerStubFactory, and interface IOpenflowController in
module ofcontroller.py, and unittests in test_ofcontroller.py. Improve
usability of Callback class in ofprotoops: allow passing args directly
to calls. Add cookie to the state of a pending operation, to verify
that a request and a reply in an operation are of matching
types. Fixed regular expression tests that match randomly generated
data.